fill again contract

+++ Models/CustomServiceItem.cs
@@ -14,7 +14,7 @@ namespace erp_backend.Models
 		public int Quantity { get; set; } = 1;
 
 		[Range(0, 100, ErrorMessage = "Thu? ph?i t? 0-100%")]
-		public decimal Tax { get; set; }
+		public float Tax { get; set; }
 
 		[Range(0, double.MaxValue, ErrorMessage = "T?ng ti?n ph?i l?n h?n ho?c b?ng 0")]
 		public decimal Total { get; set; }
